Add all the Clang libraries to the bridge
The bridge supports the whole set of the Clang's libraries
including the headers one - 'clang_headers'. Desired
libraris just can be added as dependencies to a rule using
the usual 'deps' attribute.

In order to get the shared library when the 'clang_libclang'
library is taken into account, the rule 'clang_copy_libclang'
has been added. The rule must be added to the 'data' attribute
of a target in form of '@local_llvm//:clang_copy_libclang'
where 'local_llvm' is the desired name of the 'llvm_configure'
repository rule. The shared library will appear into the
'bazel-bin/external/local_llvm' directory and may be grabbed
to run the built application.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>