Enable support for Z3 solver
When LLVM/Clang is built on an environment that supports the Z3 Solver
(https://github.com/Z3Prover/z3), the code of some LLVM and Clang
libraries is enriched by the Z3 library and projects dependent on the
libraries must be linked against Z3. To enable the linking, a new
environment variable is added: Z3_INSTALL_PREFIX. If the environment
variable is set, the Z3 library will be copied into the building
directory and a target for the Z3 Solver will be added into the 'deps'
attribute of the 'llvm_support' library. This construction reflects
what exactly CMake does for Z3-enabled builds of LLVM.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>