Add the cxx_* targets for libc++
The following targets will be added to deal with the libc++ standard
library:

 - libcxx_static
 - libcxx_shared
 - libcxx_copy_shared
 - libcxx_abi_static
 - libcxx_abi_shared
 - libcxx_copy_abi_shared

The 'libcxx' prefix can be configured in the 'WORKSPACE' file using the
'libcxx_prefix' optional parameter of the 'llvm_configure' repository
rule.

To protect adding of the targets as dependencies, a number of macroses
are put into the 'llvm_config.bzl' generated file: the `if_cxx_linked`
function that evaluated its first argument when the local LLVM
installation is linked against libc++ and 'llvm_cxx_copts' array that
contains the `-stdlib=libc++` compiler option in this case.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>