Link llvm_debug_info_pdb against DIA SDK if it is enabled
If the 'llvm_debug_info_pdb' library is linked against the DIA SDK
on the client machine, the bridge should add the SDK to the 'deps' list
of the generated 'cc_library' rule. In order to get this done, the
bridge reads the 'LLVMExports.cmake' file and looks for the
'LLVMDebugInfoPDB' library's dependencies, gets all the dependencies
that aren't started from 'LLVM' prefix and symlinks them into the build
directory for the llvm bridge.

The logic works only on Windows.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>