Add MLIR's libraries
The MLIR project has been migrated to the LLVM's monorepo; thus, it
looks as a good idea to add MLIR support to the bridge. All the 57 MLIR's
libraries have been added so that we can use them from the Bazel 'BUILD'
files.

A new default prefix, 'mlir_', was defined for the MLIR related
components.

A new header library, 'mlir_headers', was defined for the header files,
provided by MLIR. If the 'add_headers_to_deps' parameter of the
'llvm_configure' repository rule in your 'WORKSPACE' file is set to
'False', you should manually add the header library in the 'deps'
attribute of your targets that leverage MLIR. Otherwise, the header
files provided by MLIR as well as LLVM itself will be available for your
sources automatically.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>