Add an example of using tablegen
The 'llvm_print_physical_registers' example demonstrates
how to use 'llvm-tablegen' from a 'BUILD' file as a dependency
for a 'cc_binary' target. The example just parses a register
file definition for an imaginary target and prints every
register classes and the registers the class is composed of.

To run the example, do the following:

$ bazel run --action_env LLVM_INSTALL_PREFIX=<path to llvm> //:llvm_print_physical_registers

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>