Support the clang-cl compiler on Windows
The toolchain resolution mechanism is enabled and clang-cl
compiler can be used on Windows.

To enable the compiler, a config option 'clang-cl' has been
added. Bazel can automatically detect Clang or the BAZEL_LLVM
environment variable can be used (actually, the variable works
as a duplicate for the LLVM_INSTALL_PREFIX but it has a different
meaning: BAZEL_LLVM is a path to the compiler, while
LLVM_INSTALL_PREFIX is a path to LLVM libraries).

see https://docs.bazel.build/versions/master/windows.html#build-c-with-clang
for more information.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>