Add libc++ includes as the cxx_headers library
The header files from libc++ can now be used from targets
as the <prefix_of_cxx>_headers dependency.

The '_llvm_get_include_rule' function is now able to accept
the 'includes' optional parameter. It is required since
the libc++ headers take place in the 'include/c++/v1'
directory.

Since the header library is added to dependencies, there can
be some collisions between the default C++ standard library
(libstdc++, for example) and the provided libc++ headers.
To eliminate the collisions, the "-nostdinc++" compilation
option should be added to the 'copt' parameter of the targets
(wrapped into the 'if_cxx_linked' condition).

We assume the local installation of LLVM is linked against
libc++ if the latter exists there.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>