Remove any libc++ related options from examples/BUILD
Any libc++ related configurations and options, such as 'llvm_cxx_linked',
'llvm_linkopts', and 'if_cxx_linked' has been removed from the
'examples/BUILD' file: the C++ library is a part of the toolchain and must
be configured in the project level (in a .bazelrc file for example), not in
a BUILD file. The BUILD file must be clean and not depend on host and target
environments.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>