Rename llvm_cxx_linked into llvm_has_cxx
In fact, the 'llvm_cxx_linked' and 'if_cxx_linked' didn't check
whether LLVM is linked against the libc++ implementation of the
standard library, they just checked if there is a file with
the implementation ('libc++.a', for example). The variable
and function are renamed to better reflect their responsibilities.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>